Updated to release.
Updated discord.js to version 13.0.1, reset GuideBot's version numbering.

Updated the help command, added two new slash commands `leave` and `stats`, `stats` is essentially a recreation of the regular command to show a comparison between slash commands and traditional to visually tell you that they're not that different.

Eslint fixed all of the files for consistency, speaking of consistency I changed how all of the main files are exported as a couple of the files were exported slightly different.