Slash Commands (#143)
* Slash Commands

Added slash command support to the original (main) version of guidebot.

Updated discord.js to version 13.0.1, updated the throw in the index to require version 16 of nodejs.

Brought both versions of guidebot to the same functional level.

* Fixed a small mistake

The setup placed the partial array inside an array causing an error.

* Minor update

Updated the readme to be concise with the other version.
Updated the setup script, removed the deprecated ownerId field.

* Consistency fix

The prefix for the bot in two files were different compared to the one in the set up script. Fixed for consistency.

* The real quick fix

Actually fixed the damn thing this time.