2.0.0
Removed the installer and all required packages.
Removed the `loadCommand` and `unloadCommand` functions.
Removed the `wait` function, was no longer used.
Removed the `owners` functionality as the concept was too advanced for an example bot.
Required the `logger` file where it was needed instead of binding it to `client`.
Required the `settings` database and `config` file where needed.
Moved the `clean` function into the `eval` command.
Updated all of the code to reflect these changes.
Moved the discord bot token into an `.env` file for better security.
Added `owner` to the `.env` file.

To-do;
Change `levelCache` into a function to pull it off the `client` property.