Fixed Help
Fixed the help command, it will now respond if you supply an invalid command name, or alias.

It will now display the help message if an alias is supplied.

Updated the `messageCreate` event to output the error into the console.

Updated the example config to take advantage of the Nullish coalescing operator